fix: fixing completion for "after topic command" / subcommands
Autocompletion works on first level thanks to `__ltrim_colon_completions`
function that trims the autocompletion engine and return the right most part
of it.
However, on the input side, the issue persists and any after topic completion fails
as stated on https://github.com/oclif/plugin-autocomplete/issues/9.

The fix is to locally remove the colon to be processed like a completion word
separator.